requirements.txt
sentencepiece
joblib
mecab-python-windows
more_itertools
